backend/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
pydantic==2.10.6
pydantic-settings==2.7.1
openai==1.60.2
PyYAML==6.0.2
python-dateutil==2.9.0.post0
